requirements.txt
numpy>=1.26
scipy>=1.11
PyYAML>=6.0
torch>=2.5
torchvision>=0.20
torchaudio>=2.5
matplotlib>=3.8
seaborn>=0.13
pandas>=2.2
scikit-learn>=1.5
tqdm>=4.66
pytest>=8.2
